+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Ability to export run props as pandas dataframe (gorarakelyan)
 - Improve Images Explorer rendering performance through better images list virtualization (roubkar)
 
 ## 3.1.0 Nov 20 2021